requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.1
supabase==1.0.3
pydantic==1.10.11
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.0
httpx==0.24.1
